fix(catalog): keep connections adapter in backend modules
