--- .changeset/tidy-apes-connect.md
@@ -1,5 +1,0 @@
----
-'@backstage/integration': minor
----
-
-Added support for creating a GitHub credentials provider backed by the connections service.
