+++ packages/integration/src/github/DefaultGithubCredentialsProvider.ts
@@ -17,8 +17,6 @@
 import { GithubCredentials, GithubCredentialsProvider } from './types';
 import { ScmIntegrationRegistry } from '../registry';
 import { SingleInstanceGithubCredentialsProvider } from './SingleInstanceGithubCredentialsProvider';
-import type { ConnectionsService } from '@backstage/connections';
-import { GithubIntegrationConfig } from './config';
 
 /**
  * Handles the creation and caching of credentials for GitHub integrations.
@@ -43,19 +41,8 @@ export class DefaultGithubCredentialsProvider
     return new DefaultGithubCredentialsProvider(credentialsProviders);
   }
 
-  /**
-   * Creates a credentials provider backed by the connections service.
-   *
-   * @param connections - The connections service used to resolve GitHub credentials.
-   * @public
-   */
-  static fromConnections(connections: ConnectionsService) {
-    return new DefaultGithubCredentialsProvider(new Map(), connections);
-  }
-
   private constructor(
     private readonly providers: Map<string, GithubCredentialsProvider>,
-    private readonly connections?: ConnectionsService,
   ) {}
 
   /**
@@ -83,56 +70,6 @@ export class DefaultGithubCredentialsProvider
    * @returns A promise of {@link GithubCredentials}.
    */
   async getCredentials(opts: { url: string }): Promise<GithubCredentials> {
-    if (this.connections) {
-      // Ask the connections service to select auth for this URL. A host may
-      // have different GitHub Apps for different organizations, so this
-      // selection cannot be done once when the provider is created.
-      const connection = await this.connections.find({
-        type: 'github',
-        url: opts.url,
-        authMethods: ['app', 'token', 'none'],
-      });
-      const { auth } = connection;
-
-      // Keep one provider per host and selected auth method. In particular,
-      // reusing an App provider preserves its installation-token cache.
-      const providerKey = `${connection.host}:${auth.method}:${
-        auth.method === 'app' ? auth.appId : ''
-      }`;
-
-      let provider = this.providers.get(providerKey);
-      if (!provider) {
-        // Adapt the connection schema to the existing provider configuration
-        // so credential creation and token caching stay in one implementation.
-        const config: GithubIntegrationConfig = {
-          host: connection.host,
-          apiBaseUrl: connection.apiBaseUrl,
-          rawBaseUrl: connection.rawBaseUrl,
-        };
-
-        if (auth.method === 'app') {
-          config.apps = [
-            {
-              appId: Number(auth.appId),
-              privateKey: auth.privateKey,
-              clientId: auth.clientId,
-              clientSecret: auth.clientSecret,
-              webhookSecret: auth.webhookSecret,
-              publicAccess: auth.publicAccess,
-              allowedInstallationOwners: auth.orgs,
-            },
-          ];
-        } else if (auth.method === 'token') {
-          config.token = auth.token;
-        }
-
-        provider = SingleInstanceGithubCredentialsProvider.create(config);
-        this.providers.set(providerKey, provider);
-      }
-
-      return provider.getCredentials(opts);
-    }
-
     const parsed = new URL(opts.url);
     const provider = this.providers.get(parsed.host);
