+++ plugins/catalog-backend-module-github-org/src/createGithubCredentialsProvider.test.ts
@@ -1,0 +1,73 @@
+/*
+ * Copyright 2026 The Backstage Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import type { ConnectionsService } from '@backstage/connections';
+import {
+  GithubIntegrationConfig,
+  SingleInstanceGithubCredentialsProvider,
+} from '@backstage/integration';
+import { createGithubCredentialsProvider } from './createGithubCredentialsProvider';
+
+jest.mock('@backstage/integration', () => ({
+  ...jest.requireActual('@backstage/integration'),
+  SingleInstanceGithubCredentialsProvider: {
+    create: jest.fn(),
+  },
+}));
+
+describe('createGithubCredentialsProvider', () => {
+  beforeEach(() => {
+    jest.resetAllMocks();
+    (
+      SingleInstanceGithubCredentialsProvider.create as jest.MockedFunction<
+        typeof SingleInstanceGithubCredentialsProvider.create
+      >
+    ).mockImplementation((config: GithubIntegrationConfig) => ({
+      getCredentials: async () => ({
+        type: 'token',
+        token: `${config.host}-token`,
+      }),
+    }));
+  });
+
+  it('creates token providers from token connections', async () => {
+    const find = jest.fn().mockResolvedValue({
+      type: 'github',
+      title: 'GitHub Enterprise',
+      host: 'github.example.com',
+      auth: {
+        method: 'token',
+        token: 'connection-token',
+      },
+    });
+    const provider = createGithubCredentialsProvider({
+      find: find as ConnectionsService['find'],
+    });
+
+    await provider.getCredentials({
+      url: 'https://github.example.com/backstage/backstage',
+    });
+
+    expect(SingleInstanceGithubCredentialsProvider.create).toHaveBeenCalledWith(
+      {
+        host: 'github.example.com',
+        apiBaseUrl: undefined,
+        rawBaseUrl: undefined,
+        token: 'connection-token',
+      },
+    );
+  });
+});
