+++ plugins/catalog-backend-module-github/src/module/githubCatalogModule.test.ts
@@ -24,16 +24,9 @@ import {
 import { githubCatalogModule } from './githubCatalogModule';
 import { GithubLocationAnalyzer } from '../analyzers/GithubLocationAnalyzer';
 import { connectionsServiceFactory } from '@backstage/connections';
-import { DefaultGithubCredentialsProvider } from '@backstage/integration';
 
 describe('githubCatalogModule', () => {
-  afterEach(() => jest.restoreAllMocks());
-
   it('should register provider at the catalog extension point', async () => {
-    const fromConnections = jest.spyOn(
-      DefaultGithubCredentialsProvider,
-      'fromConnections',
-    );
     let addedProviders: Array<EntityProvider> | undefined;
     let usedSchedule: SchedulerServiceTaskScheduleDefinition | undefined;
 
@@ -93,6 +86,5 @@ describe('githubCatalogModule', () => {
     expect(analysisExtensionPoint.addScmLocationAnalyzer).toHaveBeenCalledWith(
       expect.any(GithubLocationAnalyzer),
     );
-    expect(fromConnections).toHaveBeenCalledTimes(1);
   });
 });
